// src/app/models/warehouse/warehouse-add-or-change-model.ts
export class WarehouseAddOrChangeModel {
    id: string;
    code: string;
    name: string;
    email: string;
    phoneNumber: string;
    addressDetail: string;
    districtName: string;
    districtId: number;
    provinceId: number;
    provinceName: string;
    roadId: number;
    roadName: string;
    status: number;
    type: number;
    description: string;
    statusName: string;
    typeName: string;
    vendorId: string;
    vendorName: string;
    villageId: number;
    villageName: string;
    lat: number;
    lng: number;
}